+++ Runtime/AvatarCreator/Scripts/WebRequests/AvatarAPIRequests.cs
@@ -1,19 +1,18 @@
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using ReadyPlayerMe.Core;
-using UnityEngine;
-using UnityEngine.Networking;
 
 namespace ReadyPlayerMe.AvatarCreator
 {
     public class AvatarAPIRequests
     {
+        private const string INVALID_AVATAR_ID_ERROR_MESSAGE = "Avatar ID is null or empty. Please provide a valid avatar ID.";
+        
         private const string RPM_AVATAR_V1_BASE_URL = Env.RPM_API_V1_BASE_URL + "avatars";
         private const string RPM_AVATAR_V2_BASE_URL = Env.RPM_API_V2_BASE_URL + "avatars";
 
@@ -110,6 +109,7 @@ namespace ReadyPlayerMe.AvatarCreator
 
         public async Task<AssetColor[]> GetAvatarColors(string avatarId, AssetType assetType = AssetType.None)
         {
+            ValidateAvatarId(avatarId);
             var colorParameters = assetType.GetColorProperty();
             if (string.IsNullOrEmpty(colorParameters))
             {
@@ -131,6 +131,7 @@ namespace ReadyPlayerMe.AvatarCreator
 
         public async Task<AvatarProperties> GetAvatarMetadata(string avatarId, bool isDraft = false)
         {
+            ValidateAvatarId(avatarId);
             var url = $"{RPM_AVATAR_V2_BASE_URL}/{avatarId}.json?";
 
             if (isDraft)
@@ -172,6 +173,7 @@ namespace ReadyPlayerMe.AvatarCreator
 
         public async Task<byte[]> GetAvatar(string avatarId, bool isPreview = false, string parameters = null)
         {
+            ValidateAvatarId(avatarId);
             var url = $"{RPM_AVATAR_V2_BASE_URL}/{avatarId}.glb?";
 
             if (!string.IsNullOrEmpty(parameters))
@@ -194,6 +196,7 @@ namespace ReadyPlayerMe.AvatarCreator
 
         public async Task<AvatarProperties> GetAvatarProperties(string avatarId)
         {
+            ValidateAvatarId(avatarId);
             var url = $"{RPM_AVATAR_V2_BASE_URL}/{avatarId}.json?";
 
             var response = await authorizedRequest.SendRequest<Response>(
@@ -212,6 +215,7 @@ namespace ReadyPlayerMe.AvatarCreator
 
         public async Task<byte[]> UpdateAvatar(string avatarId, AvatarProperties avatarProperties, string parameters = null)
         {
+            ValidateAvatarId(avatarId);
             var url = $"{RPM_AVATAR_V2_BASE_URL}/{avatarId}?responseType=glb&{parameters}";
 
             var response = await authorizedRequest.SendRequest<Response>(
@@ -229,6 +233,7 @@ namespace ReadyPlayerMe.AvatarCreator
 
         public async Task PrecompileAvatar(string avatarId, PrecompileData precompileData, string parameters = null)
         {
+            ValidateAvatarId(avatarId);
             var json = JsonConvert.SerializeObject(precompileData);
 
             var response = await authorizedRequest.SendRequest<Response>(
@@ -245,6 +250,7 @@ namespace ReadyPlayerMe.AvatarCreator
 
         public async Task<string> SaveAvatar(string avatarId)
         {
+            ValidateAvatarId(avatarId);
             var response = await authorizedRequest.SendRequest<Response>(
                 new RequestData
                 {
@@ -259,6 +265,7 @@ namespace ReadyPlayerMe.AvatarCreator
 
         public async Task DeleteAvatar(string avatarId, bool isDraft = false)
         {
+            ValidateAvatarId(avatarId);
             var url = $"{RPM_AVATAR_V2_BASE_URL}/{avatarId}/";
 
             if (isDraft)
@@ -273,6 +280,14 @@ namespace ReadyPlayerMe.AvatarCreator
                 ctx: ctx);
 
             response.ThrowIfError();
+        }
+
+        private void ValidateAvatarId(string avatarId)
+        {
+            if (string.IsNullOrEmpty(avatarId))
+            {
+                throw new Exception(INVALID_AVATAR_ID_ERROR_MESSAGE);
+            }
         }
     }
 }